Save some renaming and a new header
My new policy is that everything that should be documented with doxygen, should
be put in the include directory.  I should figure out if it is possible to
generate two different documentations.  One that documents everything and one
that documents the public interface provided in include.  I believe this should
be feasible with two different doxygen configuration files in src/doc.