Crush nasty bug in FEM routine
A copy and paste mistake turned a w into a u.

Add new macros in makefile, which code directory path.  This is important, since
on the ubuntu file system the home directory needs a kerberos ticket to work
properly.  This is not needed for the scratch directory.  On OSX I will anyway
use a different directory.

Some improvement in the documentation have been made.