Save compiling version of initial data
I have made all necessary modifications such that initial data also works.
Added a lot to the documentation.  I also use the new esfem_error.h technique to
report errors in function bodies.

I have noticed that I may not be appropriated to prefix the error string with a
message.  The problem is if you catch and throw with nested, then it may be,
that the prefix string will be displayed twice.  Perhaps I should change this in
the future.