Remove initial value mistake
I had some bugs in my code.  First, file.good() is not the same as
static_cast<bool>(file).  You have to test for !file.fail().  In addition I set
in the parameter file the gamma value to 30, but it should be zero.  Last, my
initial values was still a random distribution.