Save progress in the experiment
I was right with my suspicion.  The error were evaluated on the wrong surface.
There is still an error in the code.  My guess is the right-hand side of the
PDE.  I should definitely get that right.  Especially the w should in the next
run display the correct order of convergence.

So first, I try to get the correct EOC for u and w on the interpolated surface.
Then I tackle the surface and finally I get the velocity right.