Save again bad experiment
I have made the following frightening observation.  If I set the parameter
epsilon and delta zero, and further I do not assemble the load vector, then the
solution of the linear system should be the initial X.app itself.  But the error
is around 1e-5.  This is impossible.  I will try a different solver.

In addition I forgot the evolve the surface itself.