Save bad test
I had to put next_timeStep() above the interpolation.  Else I am comparing wrong
surfaces.  The first error is good, around 1e-8, and it decreases for finer
sphere like it should be.  However, the latter error are simply not improving.
I have no idea why this is the case.

As a side remark: I have interpolated the initial sphere and then multiply the
nodes with e^t.  The difference is very small compared with interpolating the
deformation function, around 1e-16 - 1e-18, but it should be zero!