Accept L2- and H1-error implementation
Unfortunately I could not use the debugger to inspect my vectors.  However the
results looks very plausible, because, me error became bigger.

I could observe basically the same EOC like for the incomplete L2/H1-error.

It seems that my code does not compute the correct result.  If the mesh and the
time step goes in the correct ratios down the max(L2/H1) error does not
decrease.  I believe there is still an error in the code.

Currently, I do not know where the bug is hiding.