Run some good experiments
I did for the surface logarithmic sphere the same experiments as for the
Dalquist sphere.  The results are similar.  If the alpha parameter is one, then
the resulting scheme is very good.  If the alpha parameter is off, then the
convergence is very slow; in the H1 norm it is sometimes even negative.  I
should run the alpha equal zero experiment again on the big working station for
up to sphere5.  Else people might think, that my code is not working.