Fix bug and save good experiments
I forgot to add default value for the logistic growth parameters, so at run time
the code aborted.  The EOC are good, but not as perfect as the EOC for the
surface evolution.  I do not know why.  Anyway the right-hand side seems to be
good.

Next step is to code the surface evolution part.